updating to lfs
for some reason i get an error now within the unity ecosystem. it continously reloads and searches for listed file changes. this seems to be a 3 year issue. i cant sign into my github account as well within unity. this is strange to me because it worked one one computer but not another. so this error seems to be very strange indeed. hopefully this gets patched but so far it hasnt been for several years. i will have to use desktop and unty at the same time which feels inelegant.